03-Intelligence-Memory-Service/src/memory/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
alembic>=1.13.2
redis>=5.0.1
structlog>=25.0.0
orjson>=3.10.0
uvloop>=0.19.0 ; sys_platform != "win32"
httpx>=0.24.1
prometheus-client>=0.20.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-exporter-otlp>=1.25.0
python-json-logger>=2.0.7
pydantic-settings>=2.2.1
pytest>=7.4.0
pytest-asyncio>=0.23.2
testcontainers>=4.7.1
psycopg[binary,pool]>=3.1.18
croniter>=1.4.1
supabase>=2.0.0
